BST-5328: Validate rules db imports
Added the possibility to have an imports statement in the rules db.
Added validation to make sure that each imported rules exists and no
import cycle are created (A import B, B import A).

For now, only the validation of these rules is done. Next, will be to
add  the merging into a single namespace & uploading.