BST-5993: Registry base path option (#33)
The CLI options to specify the module, scanners or rules-realm paths
were replaced by a single option: registry-path. Now the path of the
scanners and rules-realm are derived from the provided base registry.

This was made in prevision of an upcoming changes to the way updated
modules/rules are tracked.