BST-5993: Detect updated dependencies (#34)
The import & updated logic was modified to handle detection of module
whose dependencies were modified. The new logic is to first load all the
scanners & rules-realm from the registry, marking the updated namespace,
and finally resolving the rules, default rule and updated status for
each scanner module.

To avoid confusion, the shared module was split into schema, config &
models.

Integrations tests were also added in order to better e2e test the
upload. Those tests create a temporary git repo for the registry,
copies the module & rules yaml files and run cli. This setup is closer
to the real use of the cli. They are not considered "module tests" since
the request to the backend are still mocked.