BST-6070: Validate rules db integration tests
Moved all tests that uses the validate cli app directly to the
integration tests file. The tests where also reworked to use the same
pattern as for the upload: copy sample yaml files to a temporary git
repo and run the cli on them.

The use_sample utility function to setup these files was converted to a
fixture to help with reusability.