BST-6070: Validate server-side scanners
Added validation for the new server-side scanners. These scanners are
almost identical to the regular scanner, but they don't have steps, must
have `server_side: true`. Additional validation to the module schema was
added to make sure that a module without rules must be set as
server-side. Also, check that namespaces are unique across all scanners,
server-side or not.