BST-6070: Validate namespace module schema (#37)
Like what was done for the rules and rules db, the schema & validation
was converted to a pydantic model. Package `jsonschema` was also removed
since it is no longer used.

New module `errors` was created to share conversion of pydantic
validation errors. I plan to improve the overall errors handling in a
future pr.