BST-6077: Validate server-side scanners rules (#39)
Server-side scanners rules definition are now validated in the same way
the regular scanners are.

The function used to load all rules.yaml files was also slightly
refactored to only return a path rather than a struct. The only
"breaking" change is that we now log the relative path of the file we
are validating rather than the namespace. IMO, this is needed since a
namespace can now be in 3 different subfolders, so having the path makes
it easier to find which file is failing validation.